# Conflicts:
#	src/main/java/com/example/backendpensionat/Repos/ContractCustomerRepo.java

+++ src/main/java/com/example/backendpensionat/Repos/BlacklistRepo.java
@@ -3,5 +3,5 @@ package com.example.backendpensionat.Repos;
 import com.example.backendpensionat.Models.Blacklist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlacklistRepo extends JpaRepository<Blacklist, Integer> {
+public interface BlacklistRepo extends JpaRepository<Blacklist, Long> {
 }
